Refactoring pyserini.encode and pyserini.search.faiss
