+++ integrations-optional/dense/test_ance.py
@@ -21,8 +21,8 @@ import os
 import unittest
 
 from integrations.utils import clean_files, run_command, parse_score_qa, parse_score_msmarco
+from pyserini.encode import QueryEncoder
 from pyserini.search import get_topics
-from pyserini.search.faiss._searcher import QueryEncoder
 
 
 class TestAnce(unittest.TestCase):
