+++ integrations-optional/dense/test_dpr.py
@@ -22,8 +22,8 @@ import os
 import unittest
 
 from integrations.utils import clean_files, run_command, parse_score_qa
+from pyserini.encode import QueryEncoder
 from pyserini.search import get_topics
-from pyserini.search.faiss._searcher import QueryEncoder
 
 
 class TestDpr(unittest.TestCase):
