+++ integrations-optional/dense/test_tct_colbert.py
@@ -21,8 +21,8 @@ import os
 import unittest
 
 from integrations.utils import clean_files, run_command, parse_score
+from pyserini.encode import QueryEncoder
 from pyserini.search import get_topics
-from pyserini.search.faiss._searcher import QueryEncoder
 
 
 class TestTctColBert(unittest.TestCase):
