+++ pyserini/encode/__init__.py
@@ -21,6 +21,7 @@ from ._base import QueryEncoder, DocumentEncoder, JsonlCollectionIterator, Jsonl
 from ._aggretriever import AggretrieverDocumentEncoder, AggretrieverQueryEncoder
 from ._ance import AnceEncoder, AnceDocumentEncoder, AnceQueryEncoder
 from ._auto import AutoQueryEncoder, AutoDocumentEncoder
+from ._bpr import BprQueryEncoder
 from ._cached_data import CachedDataQueryEncoder
 from ._cosdpr import CosDprEncoder, CosDprDocumentEncoder, CosDprQueryEncoder
 from ._dpr import DprDocumentEncoder, DprQueryEncoder
