+++ pyserini/search/faiss/__main__.py
@@ -20,16 +20,22 @@ import os
 import numpy as np
 from tqdm import tqdm
 
+from pyserini.encode import QueryEncoder
+
+from pyserini.encode import AnceQueryEncoder
+from pyserini.encode import BprQueryEncoder
 from pyserini.encode import CosDprQueryEncoder
+from pyserini.encode import DprQueryEncoder
+from pyserini.encode import TctColBertQueryEncoder
+
 from pyserini.encode._pca import PcaEncoder
 from pyserini.output_writer import get_output_writer, OutputFormat
 from pyserini.query_iterator import get_query_iterator, TopicsFormat
 from pyserini.search.faiss._searcher import (AutoQueryEncoder, AggretrieverQueryEncoder, OpenAIQueryEncoder,
-                                             QueryEncoder, AnceQueryEncoder, BinaryDenseSearcher, BprQueryEncoder,
-                                             DprQueryEncoder, DkrrDprQueryEncoder, ClipQueryEncoder, TctColBertQueryEncoder)
+                                             DkrrDprQueryEncoder, ClipQueryEncoder)
 from pyserini.search.lucene import LuceneSearcher
 from ._prf import DenseVectorAveragePrf, DenseVectorRocchioPrf, DenseVectorAncePrf
-from ._searcher import FaissSearcher
+from pyserini.search.faiss import FaissSearcher, BinaryDenseSearcher
 
 # Fixes this error: "OMP: Error #15: Initializing libomp.a, but found libomp.dylib already initialized."
 # https://stackoverflow.com/questions/53014306/error-15-initializing-libiomp5-dylib-but-found-libiomp5-dylib-already-initial
