+++ tests-optional/test_load_encoded_queries.py
@@ -19,7 +19,7 @@
 import unittest
 
 from pyserini.search import get_topics
-from pyserini.search.faiss._searcher import QueryEncoder
+from pyserini.encode import QueryEncoder
 
 
 class TestLoadEncodedQueries(unittest.TestCase):
